# Remove chars 0-0x1F and 0x7F from tag values