I was hoping that InfluxDB's influx encoder would take care of all character encoding issues, but I should have actually read the code first. If you give it a tag value with an "invalid" character, it doesn't escape it or even remove it - it just flags it as an error and outputs nothing for that whole line.

So we need to filter tag values to remove disallowed characters - 0x00-0x1F (which includes \n and \r) plus 0x7F - before sending the tag value to the encoder.

Where this is coming up: Cisco devices have multiline strings for SysDescr.