@tdanner 2022-08-09
The map method is a little bit slower for clean strings (no characters need replacing) and a lot faster for dirty strings. If you check for the presence of invalid characters first, it is just as fast for clean strings (since it is the same code for that path) and still much faster for dirty strings, though not as fast as if you just call map first.

Without the pre-check:

```
❯ go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/kentik/ktranslate/pkg/formats/influx
BenchmarkPrepareTagValueContainsAnyClean-10     13377864                87.39 ns/op
BenchmarkPrepareTagValueContainsAnyDirty-10      2344202               507.9 ns/op
BenchmarkPrepareTagValueMapClean-10             11286318               105.2 ns/op
BenchmarkPrepareTagValueMapDirty-10              7844433               151.4 ns/op
```

With it:

```
❯ go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/kentik/ktranslate/pkg/formats/influx
BenchmarkPrepareTagValueContainsAnyClean-10     12530122                86.41 ns/op
BenchmarkPrepareTagValueContainsAnyDirty-10      2357702               507.6 ns/op
BenchmarkPrepareTagValueMapClean-10             13497147                86.24 ns/op
BenchmarkPrepareTagValueMapDirty-10              5600560               213.7 ns/op
```